% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aus_accommodation}
\alias{aus_accommodation}
\title{Australian accommodation data}
\format{
Time series of class `tsibble`
}
\source{
Australian Bureau of Statistics, Cat No 8635.0, Table 10, and Cat No 6401.0, Table 1.
}
\description{
\code{aus_accommodation} contains quarterly data on Australian tourist accommodation
from short-term non-residential accommodation with 15 or more rooms, 1998 Q1 - 2016 Q2.
The data set also contains the Australian Consumer Price Index (CPI) for the same period.
Takings are in millions of Australian dollars,
Occupancy is a percentage of rooms occupied,
CPI is an index with value 100 in 2012 Q1.
}
\examples{

aus_accommodation

}
\keyword{datasets}
