% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aus_airpassengers}
\alias{aus_airpassengers}
\title{Air Transport Passengers Australia}
\format{
Annual time series of class `tsibble`.
}
\source{
World Bank.
}
\description{
Total annual air passengers (in millions) including domestic and
international aircraft passengers of air carriers registered in Australia.
1970-2016.
}
\examples{

aus_airpassengers

}
\keyword{datasets}
