% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aus_arrivals}
\alias{aus_arrivals}
\title{International Arrivals to Australia}
\format{Quarterly time series of class `tsibble`.}
\source{
Tourism Research Australia.
}
\description{
Quarterly international arrivals to Australia from Japan, New
Zealand, UK and the US.  1981Q1 - 2012Q3.
}
\examples{

aus_arrivals

}
\keyword{datasets}
