% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aus_fertility}
\alias{aus_fertility}
\title{Australian women fertility rate}
\format{
Time series of class `tsibble`
}
\source{
Australian Bureau of Statistics. \url{https://www.abs.gov.au/statistics/people/population/births-australia/2022}
}
\description{
\code{aus_fertility} is a yearly `tsibble` with one measured variable:
\tabular{ll}{
    Rate: \tab Fertility rate (per thousand women)\cr
}
}
\details{
Each series is uniquely identified using two keys:
\tabular{ll}{
    Region: \tab Australia, states and territories\cr
    Age: \tab Age of the woman\cr
}
based on calendar year of registration data. It covers the period from 1975--2022.
}
\examples{
aus_fertility

}
\keyword{datasets}
