% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aus_inbound}
\alias{aus_inbound}
\title{Monthly short term (<1 year) visitor arrivals to Australia}
\format{
Time series of class `tsibble`
}
\source{
Tourism Research Australia
}
\description{
\code{aus_inbound} contains monthly data with one measured variable:
\tabular{ll}{
    Count: \tab Number of individuals arriving in Australia (in thousands) \cr
}
}
\details{
Each series is uniquely identified using two keys:
\tabular{ll}{
    County: \tab Country of stay/residence \cr
    Purpose: \tab Purpose of travel
}

covering the period from Jan 2005--Feb 2020.
}
\examples{
aus_inbound

}
\keyword{datasets}
