% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aus_migration}
\alias{aus_migration}
\title{Australian migration data}
\format{
Time series of class `tsibble`
}
\source{
Australian Bureau of Statistics. \url{https://www.abs.gov.au/statistics/people/population/national-state-and-territory-population/dec-2023}. Cat No. 310102.
}
\description{
Net Overseas Migration (NOM) to Australia.
}
\details{
\code{aus_migration} contains quarterly data with one measured variable:
\tabular{ll}{
    \code{NOM}:  \tab The net gain or loss of population through immigration
    to Australia and emigration from Australia\cr
}
from 1981 Q2 to 2023 Q3 for the 6 states and 2 territories of Australia, indexed by:
\tabular{ll}{
    \code{Quarter}:  \tab Year-quarter.\cr
}

NOM is based on an international traveller's duration of stay being in or out
of Australia for 12 months or more, over a 16 month period.

Each series is uniquely identified using the key:
\tabular{ll}{
    \code{State}: The state or territory.\cr
}
}
\examples{

aus_migration

}
\keyword{datasets}
