% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aus_mortality}
\alias{aus_mortality}
\title{Australian mortality data}
\format{
Time series of class `tsibble`
}
\source{
\url{https://mortality.org/Data/STMF} (Downloaded on 29th May 2024.)
}
\description{
Mortality counts and mortality rates in Australia.
}
\details{
\code{aus_mortality} is a weekly `tsibble` with one value:
\tabular{ll}{
    \code{Mortality}:  \tab Mortality count/rate\cr
}
from 2015 week 01 to 2023 week 12 for five different age groups plus the
total, categorised for sex.

Each series is uniquely identified using three keys:
\tabular{ll}{
    \code{Sex}:  \tab Sex of the individual; \code{m} - male, \code{f} - female, and
\code{b} - both sexes combined\cr
    \code{Age}:  \tab Age group of the individual\cr
    \code{Type}: \tab Whether the \code{Mortality} is a count or a rate.\cr
}

The mortality rate is defined as the number of deaths per thousand people in
Australia in each week.

The rest of the two columns of the data set are:
\tabular{ll}{
    \code{Week}:  \tab Year-week\cr
    \code{Forecast}:  \tab Equals to 1 for all years where forecasted population exposures were used to calculate the weekly death rates.\cr
}
}
\examples{

aus_mortality

}
\keyword{datasets}
