% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aus_outbound}
\alias{aus_outbound}
\title{Monthly short term (<1 year) resident departures in Australia}
\format{
Time series of class `tsibble`
}
\source{
Tourism Research Australia
}
\description{
\code{aus_outbound} contains monthly data with one measured variable:
\tabular{ll}{
    Count: \tab Number of individuals departing Australia \cr
}
}
\details{
Each series is uniquely identified using two keys:
\tabular{ll}{
    County: \tab Destination \cr
    Purpose: \tab Purpose of travel
}

covering the period from Jan 2005--Jun 2017.
}
\examples{
aus_outbound

}
\keyword{datasets}
