% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aus_vehicle_sales}
\alias{aus_vehicle_sales}
\title{Australian vehicle sales}
\format{
Time series of class `tsibble`
}
\source{
Australian Bureau of Statistics
}
\description{
The number of new motor vehicles sold in Australia.
}
\details{
\code{aus_vehicle_sales} is a monthly `tsibble` with one value:
\tabular{ll}{
    \code{Count}:  \tab The number of vehicles sold (in thousands)\cr
}
from January 1994 to December 2017 in Australia,
indexed by:
\tabular{ll}{
    \code{Month}:  \tab Year-month.\cr
}

Each series is uniquely identified using the key:
\tabular{ll}{
    \code{Type}: The type of the vehicle sold.\cr
}
}
\examples{

aus_vehicle_sales

}
\keyword{datasets}
