% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bank_calls}
\alias{bank_calls}
\title{Call volume for a large North American bank}
\format{
Time series of class `tsibble` at 5 minute intervals.
}
\source{
Rob Hyndman
}
\description{
Five-minute call volume handled on weekdays between 7:00am and 9:05pm, beginning 3 March 2003 for 164 days.
}
\examples{

bank_calls

}
\keyword{datasets}
