% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boston_marathon}
\alias{boston_marathon}
\title{Boston marathon winning times since 1897}
\format{Annual time series of class `tsibble`.}
\source{
Boston Athletic Association.
\url{https://www.baa.org/races/boston-marathon/results/champions}
}
\description{
Winning times for events at the Boston Marathon.
1897-2019.
}
\examples{

boston_marathon

}
\keyword{datasets}
