% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canadian_gas}
\alias{canadian_gas}
\title{Monthly Canadian gas production}
\format{
Monthly time series of class `tsibble`.
}
\source{
Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D., (2008)
\emph{Forecasting with exponential smoothing: the state space approach},
Springer.
}
\description{
Monthly Canadian gas production, billions of cubic metres, January 1960 -
February 2005
}
\examples{

canadian_gas

}
\references{
\url{http://www.exponentialsmoothing.net}
}
\keyword{datasets}
