% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpp3.R
\docType{package}
\name{fpp3-package}
\alias{fpp3}
\alias{fpp3-package}
\title{fpp3: Data for "Forecasting: Principles and Practice" (3rd Edition)}
\description{
All data sets required for the examples and exercises in the book "Forecasting: principles and practice" by Rob J Hyndman and George Athanasopoulos \url{https://OTexts.com/fpp3/}. All packages required to run the examples are also loaded. Additional data sets not used in the book are also included.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://pkg.robjhyndman.com/fpp3package/}
  \item \url{https://github.com/robjhyndman/fpp3package}
  \item \url{http://OTexts.com/fpp3/}
  \item Report bugs at \url{https://github.com/robjhyndman/fpp3package/issues}
}

}
\author{
\strong{Maintainer}: Rob Hyndman \email{Rob.Hyndman@monash.edu} (\href{https://orcid.org/0000-0002-2140-5352}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item George Athanasopoulos [contributor]
  \item Mitchell O'Hara-Wild [contributor]
  \item Nuwani Palihawadana [contributor]
  \item Shanika Wickramasuriya [contributor]
  \item RStudio [copyright holder]
}

}
\keyword{internal}
