% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fpp3_packages}
\alias{fpp3_packages}
\title{List all packages loaded by fpp3}
\usage{
fpp3_packages(include_self = FALSE)
}
\arguments{
\item{include_self}{Include fpp3 in the list?}
}
\value{
A character vector of package names.
}
\description{
List all packages loaded by fpp3
}
\examples{
fpp3_packages()
}
