% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{guinea_rice}
\alias{guinea_rice}
\title{Rice production (Guinea)}
\format{Annual time series of class `tsibble`.}
\source{
World Bank.
}
\description{
Total annual rice production (million metric tons) for Guinea. 1970-2011.
}
\examples{

guinea_rice

}
\keyword{datasets}
