% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{insurance}
\alias{insurance}
\title{Insurance quotations and advertising expenditure}
\format{
Monthly time series of class `tsibble`.
}
\source{
Kindly provided by Dave Reilly, Automatic Forecasting Systems.
}
\description{
Monthly quotations and monthly television advertising expenditure for a US insurance company.
January 2002 to April 2005
}
\examples{

insurance \%>\%
  ggplot(aes(x=TVadverts, y=Quotes)) + geom_point()

}
\keyword{datasets}
