% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{melb_walkers}
\alias{melb_walkers}
\title{Average daily total pedestrian count in Melbourne}
\format{
Time series of class `tsibble`
}
\source{
Melbourne Open Data Portal. \url{https://data.melbourne.vic.gov.au}
}
\description{
Daily average total pedestrian count (across different sensors) from 2019-01-01 to
2024-05-29.
}
\examples{
melb_walkers |> autoplot(Count)

}
\keyword{datasets}
