% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nsw_offences}
\alias{nsw_offences}
\title{Monthly offences in NSW}
\format{
Time series of class `tsibble`
}
\source{
NSW Bureau of Crime Statistics and Research. \url{https://www.bocsar.nsw.gov.au/Pages/bocsar_datasets/Offence.aspx}
}
\description{
\code{nsw_offences} is a monthly `tsibble` with one measured variable:
\tabular{ll}{
    Count: \tab Number of offences reported \cr
}
}
\details{
Each series is uniquely identified using one key:
\tabular{ll}{
    Type: \tab Offence type \cr
}

covering the period from Apr 1995--Dec 2023.
}
\examples{
nsw_offences

}
\keyword{datasets}
