% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ny_childcare}
\alias{ny_childcare}
\title{New York childcare data}
\format{
Time series of class `tsibble`
}
\description{
The number of employees (in thousands) in child day care services in
New York City over the period the period from January 1990 to April 2024.
}
\details{
\code{ny_childcare} is a monthly `tsibble` with two columns:
\tabular{ll}{
    \code{Month}:  \tab Year-month\cr
    \code{Count}:  \tab Number of employees.\cr
}
}
\examples{

ny_childcare

}
\keyword{datasets}
