% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prices}
\alias{prices}
\title{Price series for various commodities}
\format{
Annual time series of class `tsibble`.
}
\source{
Makridakis, Wheelwright and Hyndman (1998) *Forecasting: methods and applications*, John Wiley & Sons: New York. Chapter 9.
}
\description{
Annual prices for eggs, chicken, copper, nails, oil and wheat.
Eggs, chicken, nails, oil and copper in $US; wheat in British pounds. All prices adjusted for inflation.
}
\examples{

prices |> autoplot(wheat)

}
\keyword{datasets}
