% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{souvenirs}
\alias{souvenirs}
\title{Sales for a souvenir shop}
\format{
Monthly time series of class `tsibble`.
}
\source{
Makridakis, Wheelwright and Hyndman (1998) *Forecasting: methods and applications*, John Wiley & Sons: New York. Exercise 5.8.
}
\description{
Monthly sales for a souvenir shop on the wharf at a beach resort town in Queensland, Australia.
}
\examples{

souvenirs \%>\% autoplot(Sales)

}
\keyword{datasets}
