% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_change}
\alias{us_change}
\title{Percentage changes in economic variables in the USA.}
\format{
Time series of class `tsibble`
}
\source{
Federal Reserve Bank of St Louis.
}
\description{
\code{us_change} contains percentage changes in
quarterly personal consumption expenditure, personal disposable income,
production, savings and the unemployment rate for the US, 1970 to 2016.
Original $ values were in chained 2012 US dollars.
}
\examples{

us_change

}
\keyword{datasets}
