% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{US_employment}
\alias{US_employment}
\title{US monthly employement data}
\format{Time series of class `tsibble`}
\source{
U.S. Bureau of Labor Statistics
}
\description{
\code{US_employment} is a monthly `tsibble`
}
\examples{



}
\keyword{datasets}
