% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_gasoline}
\alias{us_gasoline}
\title{US finished motor gasoline product supplied.}
\format{Time series object of class `tsibble`.}
\source{
US Energy Information Administration.
}
\description{
Weekly data beginning Week 6, 1991, ending Week 3, 2017. Units are
"million barrels per day".
}
\examples{

us_gasoline

}
\keyword{datasets}
