% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aus_births}
\alias{aus_births}
\title{Australian births data}
\format{
Time series of class `tsibble`
}
\source{
Australian Bureau of Statistics. \url{https://www.abs.gov.au/statistics/people/population/births-australia/2022}
}
\description{
Number of births in Australia.
}
\details{
\code{aus_births} is a monthly `tsibble` with one measured variable:
\tabular{ll}{
    \code{Births}:  \tab Number of births\cr
}
from January 1975 to December 2021 for the 6 states and 2 territories
of Australia, indexed by:
\tabular{ll}{
    \code{Month}:  \tab Year-month.\cr
}
#' Each series is uniquely identified using the key:
\tabular{ll}{
    \code{State}: The state or territory.\cr
}
}
\examples{

aus_births

}
\keyword{datasets}
