% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aus_mortality}
\alias{aus_mortality}
\title{Australian mortality data}
\format{
Time series of class `tsibble`
}
\source{
\url{https://mortality.org/Data/STMF} (Downloaded on 29 May 2024)
}
\description{
Weekly death counts and mortality rates in Australia.
}
\details{
\code{aus_mortality} contains weekly data with two measured variables:
\tabular{ll}{
 \code{Deaths}: \tab Death count\cr
    \code{Mortality}:  \tab Mortality rate
}
from 2015 week 01 to 2023 week 12 for five different age groups plus the
total, categorised by sex.

Each series is uniquely identified using three keys:
\tabular{ll}{
    \code{Sex}:  \tab Sex of the individual: Male, Female, or Both\cr
    \code{Age}:  \tab Age group of the individual
}

The mortality rate is defined as the number of deaths per thousand people in
Australia in each week.
}
\examples{

aus_mortality

}
\keyword{datasets}
