% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bank_calls}
\alias{bank_calls}
\title{Call volume for a large North American commercial bank}
\format{
Time series of class `tsibble` at 5 minute intervals.
}
\source{
Jonathan Weinberg
}
\description{
Five-minute call volume handled on weekdays between 7:00am and 9:05pm, beginning 3 March and 24 October 2003 (164 days).
}
\examples{

bank_calls

}
\references{
Weinberg, Brown & Stroud (2007) "Bayesian forecasting of an inhomogeneous Poisson process with applications to call center data" \emph{Journal of the American Statistical Associiation}, 102:480, 1185-1198.
}
\keyword{datasets}
