% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daily_walkers}
\alias{daily_walkers}
\title{Average daily total pedestrian count in Melbourne}
\format{
Time series of class `tsibble`
}
\source{
Melbourne Open Data Portal
}
\description{
Daily average (across different sensors) total pedestrian count in thousands from 2019-01-01 to
2024-05-29.
}
\examples{
library(tsibble)
daily_walkers

}
\keyword{datasets}
