% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otexts_views}
\alias{otexts_views}
\title{OTexts page views}
\format{
Time series of class `tsibble`
}
\description{
Daily page views on the OTexts website.
}
\details{
\code{otexts_views} is a daily `tsibble` with two columns:
\tabular{ll}{
    \code{Date}:  \tab Date for which the page views are recorded\cr
    \code{Pageviews}:  \tab Page views on the OTexts website in thousands.\cr
}
}
\examples{

otexts_views

}
\keyword{datasets}
