% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_employment}
\alias{us_employment}
\title{US monthly employment data}
\format{
Time series of class `tsibble`
}
\source{
U.S. Bureau of Labor Statistics
}
\description{
\code{us_employment} is a monthly `tsibble` containing US employment data from January 1939
to June 2019. Each `Series_ID` represents different sectors of the economy.
}
\examples{

us_employment

}
\keyword{datasets}
